FIX bugs with age field: not saving on edit and showing wrong message on validation

+++ src/main/java/ru/kata/spring/boot_security/demo/util/AppInit.java
@@ -43,8 +43,8 @@ public class AppInit {
         allRoles.add(roleService.findByName(roleAdmin));
         allRoles.add(roleService.findByName(roleUser));
 
-        userService.saveUser(new User("Admin", "Adminov", (byte) 21, "admin@admin.ru", adminPassword, adminRole));
-        userService.saveUser(new User("User", "Userov", (byte) 45, "user@user.ru", userPassword, userRole));
-        userService.saveUser(new User("All", "Allin", (byte) 33, "all@all.ru", allPassword, allRoles));
+        userService.saveUser(new User("Admin", "Adminov", 21, "admin@admin.ru", adminPassword, adminRole));
+        userService.saveUser(new User("User", "Userov", 45, "user@user.ru", userPassword, userRole));
+        userService.saveUser(new User("All", "Allin", 33, "all@all.ru", allPassword, allRoles));
     }
 }
